+++ src/app/App.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react"
 
 import i18nManager from "./system/locale/i18n"
-import Layout from "./layout"
+import Layout from "./Layout"
 import { ReactQuery, Redux, RouteProvider } from "./providers"
 import ApplicationRouter from "./routes"
