+++ src/app/providers/index.ts
@@ -1,5 +1,5 @@
 import ReactQuery from "./react-query"
-import Redux from "./state"
-import RouteProvider from "./route"
+import Redux from "./redux"
+import RouteProvider from "./route-provider"
 
 export { ReactQuery, Redux, RouteProvider }
